Implement tuples.
A major part of this commit is a move from direct comparison of types (using ==) to using a customisable function (Type.Equals()). This was neccesary to allow tuple assignment/returning without a lot of spaghetti.

The implementation of tuples themselves make the rest of the commit. The implementation is straight forward, easilier explained by the code.