Fix inference of struct members on generic struct.
To Liam: I'm not sure about all the additional passing around of generic
parameters, you might have to prod that a little.
I took a look at the broken codegen that is the next step, and it
is caused by broken name mangling resulting in both instances of RC
being equal to the one using u8.